Allow authenticated uploads of a CSV of results.
This is to allow migration of old polls into the system.  It should be idempotent, meaning we can reupload after the polls are migrated to ensure that new polls are uploaded post release.
It uses the task-queue since we have about 9000 polls and I'm not convinced that AppEngine can write that many polls at once, whereas we don't relaly care how long the migration takes (it'll be minutes not hours)