Make the upload servlet handle slightly alrger documents.
Our complete history of poll votes is about 300k, but the FileUploadServlet starts using files in /tmp if we are bigger than 10k.
Since AppEngine doesn't support local file systems, we make the inmemory filesize bigger.
Also don't turn it into a list in memory, so we can theoretically deal with it in 50-line chunks.