Add signing of the poll submission via cryptographic hash
We take a nonce, a timestamp and the poll id and sign it with a secret key.  This prevents replay attacks (since the nonce can only be used once) and the timestamp prevents storing old values and replaying them later.
This isn't entirely secure since we can't generate and communicate the nonces via a back-channel, but it's a lot better than nothing.