Update application resolvers
The Objectify library was not available from Maven central so I've added it the Github repo and amended the resolvers accordingly.

I tried upgrading to 4.1 but the library definition changed in a way that wouldn't compile anymore due to a link in the code to the internals of the library.

I've tried to do the minimum amount of work here to get the program to compile and run locally but I haven't pushed an update.